WIP: breaks many things
This is me working out how to handle playheads.

the PlayheadManager class handles them all. Mousewheel/RPi interactions
update all playheads at once.

Currently, notes are disabled, because they rely on a single playhead which
is frankly nonsense. There is a chunky refactoring to come, and it will
make everything saner. In the meantime: time for breaking changes.